desktop/src/python/requirements.txt
# PDF Processing
# pymupdf==1.23.8  # Replaced with pdfplumber for better email/structured data extraction
pdfplumber==0.11.4  # Better extraction for legal documents, preserves emails and structured data
python-docx==1.1.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# PII Detection
presidio-analyzer[gliner]==2.2.360
presidio-anonymizer==2.2.360
spacy==3.7.2

# NLP Models (download separately)
# python -m spacy download it_core_news_lg

# LLM Integration (Optional - Task 1.11)
# llama-cpp-python==0.2.27

# Visual PII Detection (Optional - Task 1.12)
# onnxruntime==1.16.3
# transformers==4.36.0

# Gemini AI Integration (Hackathon Feature)
google-generativeai==0.8.3

# Utilities
tqdm==4.66.1
